+++ engine/src/frog/lib2d/window.hpp
@@ -82,7 +82,7 @@ namespace frog::lib2d::gx
                                 int cut_width, int cut_height,
                                 float x, float y, float tex_width, float tex_height,
                                 std::uint8_t r, std::uint8_t g, std::uint8_t b, std::uint8_t a,
-                                int pivot_x, int pivot_y, float angle,
+                                float pivot_x, float pivot_y, float angle,
                                 bool flipped = false );
 
     private:
